// Frontend/template/demo_1/src/app/citas/models/usuario.model.ts
export interface UsuarioLogin {
    
    nombre: string
    rol: string
    apellidoPaterno: string
    apellidoMaterno: string
    dni: string
    celular: number
    email: string
    fechaNaciemineto: Date
    direccion: string
    especialidad: string
    contra: string
    fechaActual: Date
    user: string
    
}